Fix Battle Dome mon ability

+++ include/battle_tower.h
@@ -34,9 +34,9 @@ struct FacilityMon
     u8 ability;
 };
 
-extern const u8 gTowerMaleFacilityClasses[30];
+extern const u16 gTowerMaleFacilityClasses[30];
 extern const u16 gTowerMaleTrainerGfxIds[30];
-extern const u8 gTowerFemaleFacilityClasses[20];
+extern const u16 gTowerFemaleFacilityClasses[20];
 extern const u16 gTowerFemaleTrainerGfxIds[20];
 extern const u16 gBattleFrontierHeldItems[];
 extern const struct FacilityMon gBattleFrontierMons[];
